story refactoring
