+++ components/banner/stories/badge-introduction.config.mjs
@@ -1,0 +1,14 @@
+export default {
+	sourcePath: '../readme.md',
+	outputName: 'banner-introduction',
+	story: {
+		title: 'Components/Banner/Introduction',
+		name: 'Introduction',
+		parameters: {
+			options: {
+				showPanel: false,
+				isToolshown: false
+			}
+		}
+	}
+};
